ci: jenkins: handle-prs: Add script to handle pull requests
This script uses the GitHub API to fetch and perform certain actions
on open pull request. The goal is to be able to either merge a PR on
the CI side, or use the CI to rebase and test the PR against the latest
base branch to ensure that merging it will not lead to regressions.

The script is making use of the GitHub API which means that it is
possible for developers to locally run it in case the CI is unavailable
for any reason provided they have a valid GitHub token.

The script is also making the following assumptions to avoid
over-engineering the solution:

1) Remote 'origin' contains the base remote repository
2) CI allows anonymous access to build results
3) Write access for the given GitHub token