ci: infra: testrunner: Switch to US IBS mirror for CI
Right now the CI and the repositories are far away meaning that
download times and other factors may impact the jobs. In order to
improve the reliability we can bring them closer by using the IBS
mirror in Provo. In the future, we could so something smarter like
detect which one is the closest mirror and use that but we are not
there yet.