Makefile: Fix 'go vet' invocation for Go 1.12
Go 1.12 deprecated the 'go tool vet' call and CI started to fail with:

 vet: invoking "go tool vet" directly is unsupported; use "go vet"

Even after changing the call, go vet still complained with:
 named files must all be in one directory

It looks like 'go vet' changed its way of working and we can't pass
a series of Go files to it as it expects packages instead. We can fix
it by simply passing the entire package directory to it.