Moved .travis.yml from skuba-update to the root
Back when skuba-update was in another repository we were doing all the CI in
Travis. This commit moves the .travis.yml file that we had in the skuba-update
directory into the root of the project. This does not bring any conflicts
because we were testing the rest of skuba only in Jenkins.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>