Fix circular dependencies in Skuba class
As Platform class uses the utils package and Skuba class
is in this package and also uses Platform class, a circular
dependency is created.

To solve this, the Skuba class is moved out of the utils package
to its own caaaspctl module. This module cannot be named skuba
because under the ci pipeline, the skuba code is checkout under the
skuba directory. This must be fixed in a future change.

Signed-off-by: Pablo Chacin <pchacin@suse.com>