Adding the --annotate-only flag
This commit introduces the --annotate-only flag to skuba-update.
This flag is used to disable automatic updates but still keep
the service running just to verify and annotate to the node if updates
are available, without installing anything.

The service can be configured with a sysconfig file
/etc/sysconfig/skuba-update. Runtime flags added in SKUBA_UPDATE_OPTIONS
are automatically used by skuba-update.service in the next run without
having to restart or reload the skuba-update.timer.

For convinience also --version and --help flags have been included.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>
Signed-off-by: David Cassany <dcassany@suse.com>