Replace password with key-based auth vmware (bsc#1136366)(bsc#1138476)
This PR is addressing the same security hardering (introduced
previously for openstack PR c9131fa) configuration.

It is considered best security practice to disable the password
authentication for systems running in production. The customer
might set a password for either root or sles but this should not
be related to ssh-ing into the cluster remotely. If they want to
allow ssh password logins they have to explicitly do it by
accepting the security consequences of their actions.

If you have password authentication enabled on your SSH service
and your root password is weak, chances you are going to get hacked
are pretty high. That's why disabling root login is one of the
oldest and most used techniques to avoid system compromise on fresh
OS installations. Even if you are using strong passwords, it's
always recommended to disable root login and use a primary SSH user
instead, so then if you need to gain root access you can do it using
su or sudo command if required.

sles and root users have empty passwords. And this can be a pretty
bad thing if you want to keep attackers out of your SSH servers.
That's why the best thing you can do is to disable remote logins for
accounts with an empty password.