Remove `skuba node reset` command
This command is not necessary to be exposed to the users. It has been
proved that it has produced some misunderstandings about its goal and
under what circumstances it should be used.

`skuba node reset` was added with the assumption that it would be
executed only on nodes that failed to `bootstrap` or `join`, so
whenever this happened, the user could execute `skuba node reset` and
retry the operation from a clean slate.

Since this goal can be done implicitly when calling to `skuba node
bootstrap` or `skuba node join`, by running a implicit `kubeadm reset`
before we attempt the `kubeadm init` or `kubeadm join` operation,
there is no need to have this explicit command exposed to the users.

There is always time to add it back in case we see it's necessary or
if users are requesting to have it.