Add into the github template a "Docs" section
When creating a PR, this shouldn't be merged unless the documentation
is also updated. Otherwise, we would have code that does not match
the docs.

Obviously, this is only true if the fixes you want to merge need a
documentation update.

In any case, it is good to ask the developer to think if this needs
a documentation update.

Doc team cannot be monitoring all the issues and PRs to this repo,
so the easiest way is that the developer submits a PR to the docs repo.

When creating the PR in this repo, it is understandable that the PR to
the docs repo may not be yet ready. This is fine. You can create a "Work
in Progress" PR to the doc team repo. This will also make them aware
that some work is in the near future for them to review, and so they can
plan accordingly.

Please let's make doc team life easier with this PR :)

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>