Upgrade: do not check container runtime version
When upgrading do not check the full container runtime version.

cri-o releases are not up to speed with Kubernetes releases, and while
we have already a new Kubernetes version to upgrade to, cri-o may be
lagging behind yet.

Since the patchlevel version drift is tolerable, do not take
explicitly into account the exact version when determining if a node
is up to date when it comes to checking the container runtime version.