Lower timeout on loadbalancer monitor (bsc#1148090)
During the upgrade, especially after the 1st master has been
upgraded, there is a time window where kube-apiserver is
restarted, and no more traffic should be routed to the load
balancer.

Decreasing the timeout forces an early ping timeout and
therefore marks the master node as INACTIVE and the request
is going to an ACTIVE master instead.

Signed-off-by: Maximilian Meister <mmeister@suse.de>