Configure loadbalancer to handle upgrades (bsc#1148090)
In an upgrade, especially having multiple master nodes it is
important to have a quick reaction time of the loadbalancer.

Marking a member (master node) of the load balancer pool as
INACTIVE needs to happen almost instantly to avoid routing
requests for a master node that restarts the apiserver during
an upgrade.

This happens mostly in scripted upgrade scenarios, when master
nodes are upgraded without pause in between.

Signed-off-by: Maximilian Meister <mmeister@suse.de>