Easy recovery from bootstrap failures bsc#1148524
The bootstrap of the first master node is made by two macro phases:

  1. Bootstrapping or core components: operations like installing missing
     packages, enabling system services, running kubeadm
  2. Define core resources on top of the kubernetes cluster

At the end of phase #1 we have a working control plane, we have even a
`kubeconfig` file. However core add-ons like network (cilium), coredns,
kured and similar are not yet deployed.

Unfortunately there's no easy way to recover from a failure that happens at
the end of phase #2.

It occurred to me several times that the AWS load balancer would have problems
seeing the "just started" api-server as active. That would cause a failure at
the early beginning of phase 2.

Recovering from this failure requires a lot of manual clean up, both inside of
the cluster folder (removing CA and other files) and on the remote node.

We can allow an easy recovery from these situation: we can keep applying
phase #2 as many times as we want because the deployment of kubernetes
resources (which is the only kind of operation done during phase #2) is
idempotent.

Note well: this issue happened to me with AWS, but this kind of glitches can
happen also with other deployments.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>