Remove MustParseSemantic wherever is not needed
There are two relevant functions when parsing Kubernetes versions:
ParseSemantic and MustParseSemantic. They are identical except that the
former returns an error and the latter panics on error.

Before this commit we were abusing of MustParseSemantic, and this in
turn blurred some error logs that we have recently faced. This commit
changes MustParseSemantic with ParseSemantic wherever possible:
whenever the version does not come strictly from the skuba source code.

Last but not least, in some functions it's hard to tell whether the
given version comes from a safe place or not (e.g. the version is in a
string format). In these cases I've kept the usage of MustParseSemantic
and I've written a documentation comment explaining the possibility of
the function panicking.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>