skuba-update: fix RPM requirements
The skuba-update binary requires the kubectl binary to be available.
This dependency is usually satisfied when skuba-update is installed via
skuba because quite some packages are installed on the node at the same
time.

However this dependency is not made explicit at the package
level.

As a result of that, installing skuba-update manually doesn't pull this
required dependency.

This commit just makes the dependency explicit.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>