CODEOWNERS: Clean up file for mandatory reviews
The CODEOWNERS file was being used as a notification method for those
who were interested in specific domains. However, it was decided to
repurpose it and make reviews from CODEOWNERS mandatory. As a result of
which, we need to clean it up in order to prevent dragging people into
reviews before they sign up for it.