Deprecate extensions apiGroup for PSP
The PodSecurityPolicy in Kubernetes 1.15 used to work
in the `extentions/v1beta1` API group. In the new
version, this API group does not exist any more.
It has been deprecated and removed. The replacement is `policy/v1beta1`.

This PR replaces the apiGroup 'extensions' with 'policy' as the
last part to make SUSE provided PSPs with Kubernetes 1.16.

It also bumps the version from `0` to `1` which is needed to make
sure that `skuba addon upgrade apply` will run before the upgrade.