Reorganize platform-related testrunner parameters
Without this patch, testrunner parameters that are specific to the
terraform configuration of the platform, such as the platform's user and
the user's ssh key, as well as the number and configuration of the nodes
to create, are located at the top level of the vars.yaml config. This is
confusing because these parameters are meant to be passed to terraform
and are not strictly meant to configure testrunner itself. This change
reorganizes those parameters and updates the documentation for them,
including adding documentation for currently undocumented parameters.