Manual update of OpenStack tf files for 0.12
These are the manual operations to perform for a terraform 0.12 upgrade
in OpenStack, after a terraform 0.12upgrade:

- Booleans should not be expressed as 0 or 1, and should be written true
  or false [1]
- All the variables are made explicit, so password has to be defined in
  tfvars
- All the dependencies have to be explicit, so we are adding a
  dependency with the floating IP association [2].

This will make sure the `terraform validate` command succeeds, and that
a CaaSP cluster can be fully initialized with the current tf files.

[1]: See https://www.terraform.io/docs/configuration/expressions.html ,
section #literal-expressions [2]: The Floating IP associations should
wait for the hosts to be up, and the connection to the hosts should wait
that the floating ip is associated.  Without this patch, the terraform
apply does not know the dependency links, and could lead to the an error
where there is no element in the collection master_wait_cloudinit or
worker_wait_cloudinit, should the apply go faster than the floating ip
association.