Better node count for tests
Currently we rely on the output of skuba to know the number of
masters and workers. This is a very flimsy way of getting those
numbers as the output can change and its not really ready for
batch processing.

Instead, lets use kubectl to get those, which gives us a more
stable way of doing so thanks to selectors and the possibility
of getting the exact output we want.

This patch moves from using skuba output to guess the number of
masters to using kubectl to get the exact number of masters. The
new function can be used to get the workers as well.